agregado pdf y modificado ejercicio1

+++ ejercicio-1.c
@@ -4,11 +4,19 @@
 #include <time.h>
 #include <signal.h>
 #include <sys/wait.h>
-#define K 5
 
 int main(void)
 {
+    int K;
+    printf("Introduzca el valor para K: ");
+    scanf("%d", &K);
+
+    printf(" ****  Opciones -> 0)Huerfano  1)Zombie   2)Morir  *** \n");
     pid_t padre_pid = getpid();
+    printf("El id del proceso padre de padres es: %d\n", padre_pid);
+    printf("\n");
+
+    // Creacion de K procesos
     for (int i = 0; i < K; i++)
     {
         pid_t son_pid = fork();
@@ -17,32 +25,44 @@ int main(void)
             break;
         }
     }
+
+    /* Para evitar problemas se evita trabajar directamente sobre
+    el proceso padre de padres*/
     if (getpid() != padre_pid)
     {
-
+        //Generacion de random para la eleccion de una opcion del proceso actual.
         srand(time(NULL) ^ (getpid() << 16));
         int r = random() % 3;
-        printf("%d\t%d\n", r, getpid());
-        //0
-        if (r == 0) // huerfano
+
+        printf("El proceso: %d Elige la opcion: %d\n", getpid(), r);
+
+        pid_t parent_pid = getppid();
+
+        if (r == 0) // OPCION 0 -> HUERFANO
         {
-            pid_t parent_pid = getppid(); // obtiene la pid del proceso padre
-            kill(parent_pid, 9);          // mata al padre
-            printf("");
-            sleep(5);
+            if (parent_pid >= padre_pid)
+            {
+                kill(parent_pid, 9);
+                sleep(5);
+            }
+            else
+            {
+                printf("OYE %d ya eres HUERFANO, no mates a tu padre (que es importante para el computador)\n", getpid());
+            }
         }
-        else if (r == 1) // zombie
+        else if (r == 1) // OPCION 1 -> ZOMBIE
         {
-            pid_t parent_pid = getppid();
             kill(parent_pid, SIGSTOP);
             exit(0);
         }
-        else // <--muere
+        else // OPCION 2 -> MUERE/ TERMINA
         {
-            sleep(5);
+            sleep(3);
             exit(0);
         }
     }
+
+    //Este Else sirve para mantener el proceso principal ejecutando.
     else
     {
         while (1)
@@ -50,6 +70,5 @@ int main(void)
             wait(NULL);
         }
     }
-
     return 0;
 }
